+++ api/pkg/fjarm/idempotency/idempotencyv1/idempotency_key_test.go
@@ -1,4 +1,4 @@
-package v1
+package idempotencyv1
 
 import (
 	idempotencypb "buf.build/gen/go/fjarm/fjarm/protocolbuffers/go/fjarm/idempotency/v1"
