--- api/pkg/fjarm/idempotency/v1/idempotency.go
@@ -1,1 +1,0 @@
-package v1
